requirements.txt
openai
llama-index
PyPDF2
PyCryptodome
gradio
whisper
bs4
python-dotenv